fixed linux build
<==== Mercurial data ====>
Branch: x_VMS_1622_appserver2_customizable_timeouts

+++ common/src/utils/network/http/httptypes.cpp
@@ -1398,7 +1398,7 @@ namespace nx_http
 
         StringType KeepAlive::toString() const
         {
-            StringType result = "timeout="+StringType::number(timeout.count());
+            StringType result = "timeout="+StringType::number((unsigned int)timeout.count());
             if (max)
                 result += ", max="+StringType::number(*max);
             return result;
